refactor gurantor code
