+++ mifosng-provider/src/main/java/org/mifosng/platform/api/GuarantorsApiResource.java
@@ -22,8 +22,10 @@ import org.mifosng.platform.api.data.EntityIdentifier;
 import org.mifosng.platform.api.data.GuarantorData;
 import org.mifosng.platform.api.infrastructure.PortfolioApiDataConversionService;
 import org.mifosng.platform.api.infrastructure.PortfolioApiJsonSerializerService;
-import org.mifosng.platform.guarantor.GuarantorReadWritePlatformService;
+import org.mifosng.platform.guarantor.GuarantorReadPlatformService;
+import org.mifosng.platform.guarantor.GuarantorWritePlatformService;
 import org.mifosng.platform.infrastructure.api.ApiParameterHelper;
+import org.mifosng.platform.security.PlatformSecurityContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
@@ -34,7 +36,10 @@ import org.springframework.stereotype.Component;
 public class GuarantorsApiResource {
 
 	@Autowired
-	private GuarantorReadWritePlatformService guarantorReadWritePlatformService;
+	private GuarantorWritePlatformService guarantorWritePlatformService;
+
+	@Autowired
+	private GuarantorReadPlatformService guarantorReadPlatformService;
 
 	@Autowired
 	private PortfolioApiDataConversionService apiDataConversionService;
@@ -48,12 +53,17 @@ public class GuarantorsApiResource {
 					"state", "zip", "country", "mobileNumber",
 					"housePhoneNumber", "comment", "dob"));
 
+	private final String SystemEntityType = "m_guarantor_external";
+
+	@Autowired
+	private PlatformSecurityContext context;
+
 	@GET
 	@Consumes({ MediaType.APPLICATION_JSON })
 	@Produces({ MediaType.APPLICATION_JSON })
 	public String retrieveGuarantorDetails(@Context final UriInfo uriInfo,
 			@PathParam("loanId") final Long loanId) {
-
+		context.authenticatedUser().validateHasReadPermission(SystemEntityType);
 		Set<String> responseParameters = ApiParameterHelper
 				.extractFieldsForResponseIfProvided(uriInfo
 						.getQueryParameters());
@@ -63,7 +73,7 @@ public class GuarantorsApiResource {
 		boolean prettyPrint = ApiParameterHelper.prettyPrint(uriInfo
 				.getQueryParameters());
 
-		GuarantorData guarantorData = guarantorReadWritePlatformService
+		GuarantorData guarantorData = guarantorReadPlatformService
 				.retrieveGuarantor(loanId);
 
 		return this.apiJsonSerializerService.serializeGuarantorDataToJson(
@@ -79,7 +89,7 @@ public class GuarantorsApiResource {
 		final GuarantorCommand command = this.apiDataConversionService
 				.convertJsonToGuarantorCommand(null, loanId, jsonRequestBody);
 
-		this.guarantorReadWritePlatformService.createGuarantor(loanId, command);
+		this.guarantorWritePlatformService.createGuarantor(loanId, command);
 
 		// returns the loan Identifier for which the guarantor was associated
 		return Response.ok().entity(new EntityIdentifier(loanId)).build();
@@ -93,7 +103,7 @@ public class GuarantorsApiResource {
 		final GuarantorCommand command = this.apiDataConversionService
 				.convertJsonToGuarantorCommand(null, loanId, jsonRequestBody);
 
-		this.guarantorReadWritePlatformService.updateGuarantor(loanId, command);
+		this.guarantorWritePlatformService.updateGuarantor(loanId, command);
 
 		return Response.ok().entity(new EntityIdentifier(loanId)).build();
 	}
@@ -102,7 +112,7 @@ public class GuarantorsApiResource {
 	@Consumes({ MediaType.APPLICATION_JSON })
 	@Produces({ MediaType.APPLICATION_JSON })
 	public Response deleteGuarantor(@PathParam("loanId") final Long loanId) {
-		this.guarantorReadWritePlatformService.removeGuarantor(loanId);
+		this.guarantorWritePlatformService.removeGuarantor(loanId);
 		return Response.ok(new EntityIdentifier(loanId)).build();
 	}
 }
