+++ mifosng-provider/src/main/java/org/mifosng/platform/guarantor/GuarantorReadPlatformService.java
@@ -1,0 +1,13 @@
+package org.mifosng.platform.guarantor;
+
+import org.mifosng.platform.api.data.GuarantorData;
+
+public interface GuarantorReadPlatformService {
+
+	GuarantorData retrieveGuarantor(Long loanId);
+
+	boolean existsGuarantor(Long loanId);
+
+	GuarantorData getExternalGuarantor(Long loanId);
+
+}
