+++ mifosng-provider/src/main/java/org/mifosng/platform/guarantor/GuarantorWritePlatformServiceImpl.java
@@ -8,15 +8,11 @@ import java.util.Map;
 import java.util.Set;
 
 import org.apache.commons.lang.StringUtils;
-import org.joda.time.LocalDate;
 import org.mifosng.platform.api.commands.GuarantorCommand;
 import org.mifosng.platform.api.data.ApiParameterError;
-import org.mifosng.platform.api.data.GenericResultsetData;
-import org.mifosng.platform.api.data.GuarantorData;
 import org.mifosng.platform.client.domain.Client;
 import org.mifosng.platform.client.domain.ClientRepository;
 import org.mifosng.platform.exceptions.ClientNotFoundException;
-import org.mifosng.platform.exceptions.GuarantorNotFoundException;
 import org.mifosng.platform.exceptions.InvalidGuarantorException;
 import org.mifosng.platform.exceptions.LoanNotFoundException;
 import org.mifosng.platform.exceptions.PlatformApiDataValidationException;
@@ -24,71 +20,32 @@ import org.mifosng.platform.loan.domain.Loan;
 import org.mifosng.platform.loan.domain.LoanRepository;
 import org.mifosng.platform.noncore.ReadWriteNonCoreDataService;
 import org.mifosng.platform.organisation.domain.OfficeRepository;
-import org.mifosng.platform.security.PlatformSecurityContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class GuarantorReadWritePlatformServiceImpl implements
-		GuarantorReadWritePlatformService {
+public class GuarantorWritePlatformServiceImpl implements
+		GuarantorWritePlatformService {
 
-	private final PlatformSecurityContext context;
 	private final LoanRepository loanRepository;
 	private final ClientRepository clientRepository;
 	private final ReadWriteNonCoreDataService readWriteNonCoreDataService;
+	private final GuarantorReadPlatformService guarantorReadPlatformService;
 	// Table for storing external Guarantor Details
 	public static final String EXTERNAL_GUARANTOR_TABLE_NAME = "m_guarantor_external";
 
 	@Autowired
-	public GuarantorReadWritePlatformServiceImpl(
-			final PlatformSecurityContext context,
+	public GuarantorWritePlatformServiceImpl(
 			final LoanRepository loanRepository,
 			final ReadWriteNonCoreDataService readWriteNonCoreDataService,
 			final OfficeRepository officeRepository,
-			final ClientRepository clientRepository) {
-		this.context = context;
+			final ClientRepository clientRepository,
+			final GuarantorReadPlatformService guarantorReadPlatformService) {
 		this.loanRepository = loanRepository;
 		this.readWriteNonCoreDataService = readWriteNonCoreDataService;
 		this.clientRepository = clientRepository;
-	}
-
-	@Override
-	public boolean existsGuarantor(Long loanId) {
-		Loan loan = validateLoanExists(loanId);
-		// return if internal guarantor exists
-		if (loan.getGuarantor() != null) {
-			return true;
-		}
-		// return if an external guarantor exists
-		if (null != getExternalGuarantor(loanId)) {
-			return true;
-		}
-		// else no guarantor exists
-		return false;
-	}
-
-	@Override
-	public GuarantorData retrieveGuarantor(Long loanId) {
-		context.authenticatedUser();
-		GuarantorData guarantorData = null;
-		Loan loan = validateLoanExists(loanId);
-		// does an internal guarantor exist
-		if (loan.getGuarantor() != null) {
-			Client guarantor = loan.getGuarantor();
-			LocalDate localDate = new LocalDate(guarantor.getJoiningDate());
-			guarantorData = new GuarantorData(guarantor.getId(),
-					guarantor.getFirstName(), guarantor.getLastName(),
-					guarantor.getExternalId(), guarantor.getOffice().getName(),
-					localDate);
-		} else {
-			guarantorData = getExternalGuarantor(loanId);
-		}
-		// throw error if guarantor does not exist
-		if (guarantorData == null) {
-			throw new GuarantorNotFoundException(loanId);
-		}
-		return guarantorData;
+		this.guarantorReadPlatformService = guarantorReadPlatformService;
 	}
 
 	@Override
@@ -194,7 +151,7 @@ public class GuarantorReadWritePlatformServiceImpl implements
 			/*** update or create a new external Guarantor entry ***/
 			modifiedParametersMap.put("locale", command.getLocale());
 			modifiedParametersMap.put("dateFormat", command.getDateFormat());
-			if (getExternalGuarantor(loanId) != null) {
+			if (guarantorReadPlatformService.getExternalGuarantor(loanId) != null) {
 				readWriteNonCoreDataService.updateDatatableEntryOnetoOne(
 						EXTERNAL_GUARANTOR_TABLE_NAME, loanId,
 						modifiedParametersMap);
@@ -222,41 +179,4 @@ public class GuarantorReadWritePlatformServiceImpl implements
 		}
 		return loan;
 	}
-
-	/**
-	 * @param loanId
-	 * @return
-	 */
-	private GuarantorData getExternalGuarantor(Long loanId) {
-		GenericResultsetData genericResultDataSet = readWriteNonCoreDataService
-				.retrieveDataTableGenericResultSet(
-						EXTERNAL_GUARANTOR_TABLE_NAME, loanId, null, null);
-		if (genericResultDataSet.getData().size() == 1) {
-			List<String> guarantorRow = genericResultDataSet.getData().get(0)
-					.getRow();
-			String firstname = guarantorRow.get(1);
-			String lastname = guarantorRow.get(2);
-
-			LocalDate dateOfBirth = null;
-			if (!StringUtils.isBlank(guarantorRow.get(3))) {
-				dateOfBirth = new LocalDate(guarantorRow.get(3));
-			}
-			String addressLine1 = guarantorRow.get(4);
-			String addressLine2 = guarantorRow.get(5);
-			String city = guarantorRow.get(6);
-			String state = guarantorRow.get(7);
-			String country = guarantorRow.get(8);
-			String zip = guarantorRow.get(9);
-			String housePhoneNumber = guarantorRow.get(10);
-			String mobileNumber = guarantorRow.get(11);
-			String comment = guarantorRow.get(12);
-			GuarantorData guarantorData = new GuarantorData(firstname,
-					lastname, dateOfBirth, addressLine1, addressLine2, city,
-					state, zip, country, mobileNumber, housePhoneNumber,
-					comment);
-			return guarantorData;
-		}
-		return null;
-	}
-
 }
